Promoter test: refactor integration tests
Refactor integration tests to match the new staging environment setup format.

+++ ci-scripts/dlrnapi_promoter/tests/staging-setup/staging_environment.py
@@ -41,6 +41,7 @@ import yaml
 
 from dlrn import db as dlrn_db
 from dlrn import utils
+from string import Template
 
 
 def get_full_hash(commit_hash, distro_hash):
@@ -140,8 +141,7 @@ class StagedHash(object):
         self.full_hash = get_full_hash(self.commit_hash, self.distro_hash)
         self.repo_path = "{}/{}/{}".format(
             commit_hash[:2], commit_hash[2:2], self.full_hash)
-        self.overcloud_images_base_dir = os.path.join(
-            self.config['root-dir'], self.config['overcloud_images_base_dir'])
+        self.overcloud_images_base_dir = self.config['overcloud_images']['base_dir']
         self.docker_client = docker.from_env()
 
     def setup_images(self):
@@ -159,13 +159,16 @@ class StagedHash(object):
                                   "{}-image.tar.gz".format(self.full_hash))
         self.images_dirs[distro] = image_dir
 
+        if self.config['dry-run']:
+            return
+
         os.mkdir(os.path.join(image_dir, self.full_hash))
         # This emulates a "touch" command
         self.log.info("Create empty image in %s", image_path)
         with open(image_path, 'w'):
             pass
 
-    def promote_overcloud_images(self, candidate_name):
+    def promote_overcloud_images(self, promotion_target):
         """
         This function just creates a link to the images directory
         to emulate the existing links that need to be shifted when the real
@@ -174,11 +177,15 @@ class StagedHash(object):
         distro = self.config['distro']
         target = os.path.join(self.images_dirs[distro], self.full_hash)
         link = os.path.join(
-            self.images_dirs[distro], self.config['candidate_name'])
-        self.log.info("Link %s to %s as it was promoted to %s", target,
-                      link, self.config['candidate_name'])
+            self.images_dirs[distro], promotion_target)
+
+        if self.config['dry-run']:
+            return
+
         try:
             os.symlink(target, link)
+            self.log.info("Link %s to %s as it was promoted to %s", target,
+                          link, promotion_target)
         except OSError:
             self.log.info("Overcloud images already promoted, not creating")
 
@@ -190,7 +197,8 @@ class StagedHash(object):
         """
 
         base_image = BaseImage("promotion-stage-base:v1")
-        source_image = base_image.build()
+        if not self.config['dry-run']:
+            source_image = base_image.build()
 
         source_registry = None
         for registry in self.config['registries']:
@@ -224,14 +232,22 @@ class StagedHash(object):
                 # to emulate real life scenario
                 if ("ppc64le" in tag and image_name == suffixes[-1]):
                     continue
-                source_image.tag(full_image, tag=tag)
-                self.docker_client.images.push(full_image, tag=tag)
+                if not self.config['dry-run']:
+                    source_image.tag(full_image, tag=tag)
                 image_tag = "{}:{}".format(full_image, tag)
-                self.docker_client.images.remove(image_tag)
+
                 pushed_images.append("{}:{}".format(image, tag))
 
+                if self.config['dry-run']:
+                    continue
+
+                self.docker_client.images.push(full_image, tag=tag)
+                self.docker_client.images.remove(image_tag)
+
         self.config['results']['containers'] = pushed_images
-        base_image.remove()
+
+        if not self.config['dry-run']:
+            base_image.remove()
 
     def setup_repo_path(self):
         """
@@ -284,35 +300,30 @@ class StagedEnvironment(object):
     def __init__(self, config):
         self.config = config
 
-        self.overcloud_images_base_dir = os.path.join(
-            self.config['root-dir'], self.config['overcloud_images_base_dir'])
+        self.overcloud_images_base_dir = self.config['overcloud_images']['base_dir']
         self.stages = {}
         self.registries = {}
         self.fixture_file = self.config['db_fixtures']
         with open(self.fixture_file) as ff:
             self.fixture = yaml.safe_load(ff)
 
-        self.config['results']['commits'] = []
-        for commit in self.fixture["commits"]:
+        self.analyze_commits(self.fixture)
+
+        for commit in self.config['results']['commits']:
             stage = StagedHash(
                 self.config, commit["commit_hash"], commit["distro_hash"])
             self.stages[stage.full_hash] = stage
-            full_hash = get_full_hash(commit['commit_hash'],
-                                      commit['distro_hash'])
-            self.config['results']['commits'].append(full_hash)
-
-        self.overcloud_images_base_dir = os.path.join(
-            self.config['root-dir'], self.config['overcloud_images_base_dir'])
 
         self.docker_client = docker.from_env()
 
-    def promote_overcloud_images(self, full_hash, candidate_name):
+    def promote_overcloud_images(self):
         """
         Creates the links for the images hierarchy
         TODO: create the previous-* links
         """
-        staged_hash = self.stages[full_hash]
-        staged_hash.promote_overcloud_images(candidate_name)
+        for _, promotion in self.config['results']['promotions'].items():
+            staged_hash = self.stages[promotion['full_hash']]
+            staged_hash.promote_overcloud_images(promotion['name'])
 
     def inject_dlrn_fixtures(self):
         """
@@ -321,12 +332,17 @@ class StagedEnvironment(object):
         """
         session = dlrn_db.getSession(
             "sqlite:///%s" % self.config['db_filepath'])
+        db_filepath = self.config['db_filepath']
+        self.config['results']['inject-dlrn-fixtures'] = db_filepath
+        self.config['results']['dlrn_host'] = self.config['dlrn_host']
+
+        if self.config['dry-run']:
+            return
+
         try:
             utils.loadYAML(session, self.config['db_fixtures'])
         except sqlite3.IntegrityError:
             self.log.info("DB is not empty, not injecting fixtures")
-        db_filepath = self.config['db_filepath']
-        self.config['results']['inject-dlrn-fixtures'] = db_filepath
 
     def generate_pattern_file(self):
         """
@@ -340,6 +356,11 @@ class StagedEnvironment(object):
         """
         image_names = self.config['containers']['images-suffix']
         pattern_file_path = self.config['containers']['pattern_file_path']
+        self.config['results']['pattern_file_path'] = pattern_file_path
+
+        if self.config['dry-run']:
+            return
+
         with open(pattern_file_path, "w") as pattern_file:
             for image_name in image_names:
                 line = ("^{}$\n".format(image_name))
@@ -350,9 +371,6 @@ class StagedEnvironment(object):
         for registry_conf in self.config['registries']:
             if registry_conf['type'] == "source" and "source" in results:
                 continue
-            registry = Registry(registry_conf['name'],
-                                port=registry_conf['port'])
-            registry.run()
             if registry_conf['type'] == "source":
                 results.update({
                     'source': {
@@ -373,6 +391,12 @@ class StagedEnvironment(object):
                     'username': 'unused',
                     'password': 'unused',
                 })
+            if self.config['dry-run']:
+                continue
+
+            registry = Registry(registry_conf['name'],
+                                port=registry_conf['port'])
+            registry.run()
 
         self.config['results']['registries'] = results
 
@@ -386,6 +410,19 @@ class StagedEnvironment(object):
         Orchestrates the setting up of the environment
         """
 
+        self.config['results']['release'] = self.config['release']
+        self.config['results']['distro'] = self.config['distro']
+        self.config['results']['distro_version'] = self.config['distro_version']
+        self.config['results']['promotion_target'] = self.config['promotion_target']
+
+        template = Template(self.config['logfile_template'])
+        logfile = template.substitute({
+            'distro': self.config['distro'],
+            'distro_version': self.config['distro_version'],
+            'promoter_user': self.config['promoter_user'],
+            'release': self.config['release'],
+        })
+        self.config['results']['logfile'] = logfile
         if (self.config['components'] == "all"
            or "registries" in self.config['components']):
             self.setup_registries()
@@ -411,10 +448,19 @@ class StagedEnvironment(object):
             distro = self.config['distro']
             image_dir = os.path.join(self.overcloud_images_base_dir, distro,
                                      self.config['release'], "rdo_trunk")
-            self.log.info("Creating image dir %s", image_dir)
-            os.makedirs(image_dir)
-            self.config['results']['overcloud_images'] = \
+            self.config['results']['overcloud_images'] = {}
+            self.config['results']['overcloud_images']['base_dir'] = \
                 self.overcloud_images_base_dir
+            self.config['results']['overcloud_images']['host'] = "localhost"
+            self.config['results']['overcloud_images']['user'] = \
+                self.config['promoter_user']
+
+            self.config['results']['overcloud_images']['key_path'] = "unknown"
+
+
+            if not self.config['dry-run']:
+                self.log.info("Creating image dir %s", image_dir)
+                os.makedirs(image_dir)
 
         # Use the dlrn hashes defined in the fixtures to setup all
         # the needed component per-hash
@@ -423,17 +469,40 @@ class StagedEnvironment(object):
 
         if (self.config['components'] == "all"
            or "overcloud-images" in self.config['components']):
-            # The first commit in the fixture will be the one faking
-            # a tripleo-ci-config promotion
-            candidate_commit = self.fixture['commits'][0]
-            candidate_full_hash = get_full_hash(
-                candidate_commit['commit_hash'],
-                candidate_commit['distro_hash'])
-            self.promote_overcloud_images(
-                candidate_full_hash, self.config['candidate_name'])
+            self.promote_overcloud_images()
 
         with open(self.config['stage-info-path'], "w") as stage_info:
             stage_info.write(yaml.dump(self.config['results']))
+
+    def analyze_commits(self, fixture_data):
+        commits = []
+        for db_commit in fixture_data['commits']:
+            commit = {
+                'commit_hash': db_commit['commit_hash'],
+                'distro_hash': db_commit['distro_hash'],
+                'full_hash': get_full_hash(db_commit['commit_hash'],
+                                           db_commit['distro_hash']),
+            }
+            # Find name for commit in promotions if exists
+            for promotion in fixture_data['promotions']:
+                if promotion['commit_id'] == db_commit['id']:
+                    commit['name'] = promotion['promotion_name']
+            commits.append(commit)
+
+
+        self.config['results']['commits'] = commits
+        # First commit is currently promoted
+        currently_promoted = commits[0]
+        # Second commit is currently promoted
+        previously_promoted = commits[1]
+        # Last commit is the promotion candidate
+        promotion_candidate = commits[-1]
+
+        self.config['results']['promotions'] = {
+            'currently_promoted': currently_promoted,
+            'previously_promoted': previously_promoted,
+            'promotion_candidate': promotion_candidate,
+        }
 
     def teardown(self):
         with open(self.config['stage-info-path'], "r") as stage_info:
@@ -455,7 +524,7 @@ class StagedEnvironment(object):
 
         if (self.config['components'] == "all"
            or "overcloud-images" in self.config['components']):
-            directory = self.config['root-dir']
+            directory = self.config['overcloud_images']['base_dir']
             try:
                 self.log.debug("removing %s", directory)
                 shutil.rmtree(directory)
@@ -483,7 +552,7 @@ class StagedEnvironment(object):
             self.log.info("remote cleanup is not implemented")
 
 
-def load_config(components, db_filepath=None):
+def load_config(overrides, db_filepath=None):
     """
     This loads the yaml configuration file containing information on paths
     and distributions to stage
@@ -502,8 +571,11 @@ def load_config(components, db_filepath=None):
     config['db_fixtures'] = os.path.join(
         base_path, "fixtures", "scenario-1.yaml")
 
-    config['components'] = components
     config['results'] = {}
+
+    pprint.pprint(overrides)
+    config.update(overrides)
+    pprint.pprint(config)
 
     if (config['components'] == "all"
        or "inject-dlrn-fixtures" in config['components']):
@@ -530,12 +602,24 @@ def main():
         "registries",
     ]
     parser.add_argument('--components', default="all",
-                        help=",".join(components))
+                        help="Select components to create,".join(components))
+    parser.add_argument('--dry-run', action='store_true', default=False,
+                        help="Don't do anything, still create stage-info")
+    parser.add_argument('--promoter-user', default=os.environ.get("USER",
+                                                                  "centos"),
+                        help="The promoter user")
     args = parser.parse_args()
 
-    config = load_config(args.components)
+    overrides = {
+        "components": args.components,
+        "stage-info-path": "/tmp/stage-info.yaml",
+        "dry-run": args.dry_run,
+        "promoter_user": args.promoter_user
+    }
+    config = load_config(overrides)
 
-    config['stage-info-path'] = "/tmp/stage-info.yaml"
+    # Cli argument overrides
+
     staged_env = StagedEnvironment(config)
     if args.action == 'setup':
         staged_env.setup()
